Update woocommerce-bitcoin-price-display.php
So I really don't know what I'm doing but chatgpt got me  close to having an example feature...hopefully you can consider the update and build this into your already awesome plugin. 

As of now, the code adds the amount but only in the store's value, not in sats. Also I believe it does not add the amount if the user enters an amount using a comma (US) or decimal (EU) when adding over 1k sats. 

Anyway, this is my attempt at creating a PR but will need a lot of cleanup (probably). I should stick to coffee maybe! 😂

Ultimately, it would be awesome to put these as variables in the plugin setup page, so you could add other specific causes (opensats, nostr devs, etc.) from within wordpress, and they would appear on the checkout page. 

Lastly, I don't know if it shows up as a line item in the order back to the shop owner or not, but would be another request if possible. Thanks!